refact: refactor module
